# apps/api/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TripCircle API",
    "description": "Comprehensive API documentation for the TripCircle travel platform.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/api/posts/": {
      "get": {
        "tags": [
          "Social"
        ],
        "summary": "Get explore feed",
        "description": "Fetches a list of public trip posts with optional filters.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/posts/search": {
      "get": {
        "tags": [
          "Social"
        ],
        "summary": "Search posts",
        "description": "Searches for posts via trip titles, tags, and keywords.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "tags": [
          "Social"
        ],
        "summary": "Get a single post by ID",
        "description": "Fetches a specific post by its ID, including associated user and trip data.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/posts/trip/{tripId}": {
      "get": {
        "tags": [
          "Social"
        ],
        "summary": "Get post by trip ID",
        "description": "Fetches a post associated with a specific trip ID.",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/posts/{id}/like": {
      "post": {
        "tags": [
          "Social"
        ],
        "summary": "Toggle like on a post",
        "description": "Allows a user to like or unlike a post. The request body should contain the userId.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{id}/comments": {
      "post": {
        "tags": [
          "Social"
        ],
        "summary": "Add a comment to a post",
        "description": "Allows a user to add a comment to a post. The request body should contain the userId and commentText.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  },
                  "commentText": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{id}/comments/{commentId}": {
      "delete": {
        "tags": [
          "Social"
        ],
        "summary": "Delete a comment from a post",
        "description": "Allows a user to delete their own comment from a post. The request body should contain the userId for authorization.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/geo/route": {
      "post": {
        "tags": [
          "External Services"
        ],
        "summary": "Calculate Map Route",
        "description": "Generates routing data for a sequence of stops.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stops": {
                    "example": "any"
                  },
                  "mode": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}